lists/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.10
google-auth>=2.20
requests>=2.31
